Don't use provider aliases in vpn-data-aws.
The repo runs 'terraform validate' in each module, which doesn't
work with provider configuration_aliases:
https://github.com/hashicorp/terraform/issues/28490

The alias was only used to raise visibility on the need to pass
a provider configuration with the same region as the VPN, which
may not be the same region where the Terraform run is creating
resources. It's not a functional requirement, so it was removed.
The module will rely on documentation to handle the region requirement.